+++ src/main/java/lotto/domain/machine/LottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.machine;
 
 import java.util.List;
