+++ src/main/java/lotto/domain/machine/RandomLottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.machine;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
