+++ src/main/java/lotto/domain/winning/Rank.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.winning;
 
 public enum Rank {
     FIRST(6, 2_000_000_000, false, "6개 일치"),
