+++ src/main/java/lotto/domain/winning/WinningStatistics.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.winning;
 
 import java.util.HashMap;
 import java.util.Map;
